Адаптация клиента
