+++ Zilon.Client/Assets/Zilon/Scripts/Services/DiseaseLocalizationHelper.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e63ed85e42bac0e48844c5dccf288c1b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
