+++ Zilon.Core/Zilon.Core/MapGenerators/DiseaseGenerator.cs
@@ -86,129 +86,4 @@ namespace Zilon.Core.MapGenerators
             return nameFound;
         }
     }
-
-    public static class DiseaseNames
-    {
-        public static ILocalizedString[] Primary
-        {
-            get
-            {
-                return new[] {
-                    new LocalizedString{
-                        Ru = "Грипп"
-                    },
-
-                    new LocalizedString{
-                        Ru = "Пневмония"
-                    },
-
-                    new LocalizedString{
-                        Ru = "Хворь"
-                    },
-
-                    new LocalizedString{
-                        Ru = "Лихорадка"
-                    },
-
-                    new LocalizedString{
-                        Ru = "Болезнь"
-                    },
-
-                    new LocalizedString{
-                        Ru = "Заражение"
-                    },
-
-                    new LocalizedString{
-                        Ru = "Язва"
-                    },
-
-                    new LocalizedString{
-                        Ru = "Недостаточность"
-                    },
-
-                    new LocalizedString{
-                        Ru = "Инфекция"
-                    },
-
-                    new LocalizedString{
-                        Ru = "Помутнение"
-                    }
-                };
-            }
-        }
-
-        public static ILocalizedString[] PrimaryPreffix
-        {
-            get =>
-                new[]{
-                    new LocalizedString{
-                        Ru = "Некро"
-                    },
-                    new LocalizedString{
-                        Ru = "Гастро"
-                    },
-                    new LocalizedString{
-                        Ru = "Гипер"
-                    },
-                    new LocalizedString{
-                        Ru = "Макро"
-                    },
-                    new LocalizedString{
-                        Ru = "Прото"
-                    },
-                    new LocalizedString{
-                        Ru = "Сверх"
-                    }
-                };
-        }
-
-        public static ILocalizedString[] Secondary { 
-            get =>new[] { 
-                new LocalizedString{ 
-                    Ru = "Атипичный"
-                },
-                new LocalizedString{
-                    Ru = "Черный"
-                },
-                new LocalizedString{
-                    Ru = "Красный"
-                },
-                new LocalizedString{
-                    Ru = "Белый"
-                },
-                new LocalizedString{
-                    Ru = "Желчный"
-                },
-                new LocalizedString{
-                    Ru = "Бубонный"
-                },
-                new LocalizedString{ 
-                    Ru = "Общий"
-                },
-                new LocalizedString{ 
-                    Ru = "Кишечный"
-                }
-            };
-        }
-
-        public static ILocalizedString[] Subject {
-            get => new[] {
-                new LocalizedString{
-                    Ru = "Смерти"
-                },
-                new LocalizedString{
-                    Ru = "Крови"
-                },
-                new LocalizedString{
-                    Ru = "Печени"
-                }
-            };
-        }
-    }
-
-    public class LocalizedString : ILocalizedString
-    {
-        public string En { get; set; }
-        public string Ru { get; set; }
-    }
 }
