+++ Zilon.Core/Zilon.Core/MapGenerators/LocalizedString.cs
@@ -1,0 +1,10 @@
+﻿using Zilon.Core.Localization;
+
+namespace Zilon.Core.MapGenerators
+{
+    public class LocalizedString : ILocalizedString
+    {
+        public string En { get; set; }
+        public string Ru { get; set; }
+    }
+}
